Modifique variables, hice el config, hice un port a una bd de XAMPP

+++ src/main/java/cl/bibliotecaam/donacion/microserviciodonacion/dto/DonacionResponseDTO.java
@@ -18,5 +18,5 @@ public class DonacionResponseDTO {
     private String appaterno_dona;
     private String apmaterno_dona;
 
-    private Integer numrun_emp;
+    private Long numrun_emp;
 }
